+++ src/main/java/de/samply/app/ProjectManagerConst.java
@@ -78,6 +78,8 @@ public class ProjectManagerConst {
     public final static String FETCH_USERS_FOR_AUTOCOMPLETE_ACTION = "FETCH_USERS_FOR_AUTOCOMPLETE";
     public final static String FETCH_PROJECT_USERS_ACTION = "FETCH_PROJECT_USERS";
     public final static String FETCH_PROJECT_CONFIGURATIONS_ACTION = "FETCH_PROJECT_CONFIGURATIONS";
+    public final static String EXISTS_AUTHENTICATION_SCRIPT_ACTION = "EXISTS_AUTHENTICATION_SCRIPT";
+
 
 
     // REST Services
@@ -136,6 +138,7 @@ public class ProjectManagerConst {
     public final static String SAVE_QUERY_IN_BRIDGEHEAD = "/save-query-in-bridgehead";
     public final static String SAVE_AND_EXECUTE_QUERY_IN_BRIDGEHEAD = "/save-and-execute-query-in-bridgehead";
     public final static String DOWNLOAD_AUTHENTICATION_SCRIPT = "/download-authentication-script";
+    public final static String EXISTS_AUTHENTICATION_SCRIPT = "/exists-authentication-script";
     public final static String FETCH_DATASHIELD_STATUS = "/datashield-status";
     public final static String FETCH_PUBLICATIONS = "/publications";
     public final static String FETCH_OTHER_DOCUMENTS = "/other-documents";